# src/main/java/ismin/cours/book/model/Book.java
package ismin.cours.book.model;

import java.time.LocalDate;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Data;

@Data
@Entity
@Table(name = "BOOK")
public class Book {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    private long isbn;
    private String writer;
    private String title;

    @Column(name = "PUBLISHING_DATE")
    private LocalDate publishingDate;
}